python/discordpy/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
black==22.8.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
discord.py==2.0.1
Flask==2.2.2
frozenlist==1.3.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycparser==2.21
PyNaCl==1.5.0
tomli==2.0.1
waitress==2.1.2
Werkzeug==2.2.2
yarl==1.8.1
